Moved interfaces and base objects to Common.Client
